+++ components/select/Select.js
@@ -64,8 +64,11 @@ const InternalSelect = (props) => {
       _.uniqBy(cacheSelectItem.concat(dropdownItems), transKeys(fieldNames, 'id')),
       transKeys(fieldNames, 'id')
     )
-    // 在异步多选的时候时候才需要进行值的记录
-    dataSource && type === 'multiple' && setCacheSelectItem(selectedItems)
+    if (dataSource) {
+      // 在异步多选的时候时候才需要进行值的记录
+      type === 'multiple' && setCacheSelectItem(selectedItems)
+      autoload && remoteSearch()
+    }
   }, [])
   useEffect(() => {
     setSearchable(dataSource ? true : propsSearchable)
@@ -234,12 +237,6 @@ const InternalSelect = (props) => {
     },
     [dataSource, searchable, keyword, filterOption]
   )
-
-  useEffect(() => {
-    if (dataSource && autoload) {
-      remoteSearch()
-    }
-  })
 
   const remoteSearch = useCallback(
     (keyword) => {